Move to_native_string to _internal_utils.py to avoid circular dependency
